+++ pipelinewise/cli/__init__.py
@@ -40,7 +40,8 @@ COMMANDS = [
     'import_config',  # This is for backward compatibility; use 'import' instead
     'validate',
     'encrypt_string',
-    'partial_sync_table'
+    'partial_sync_table',
+    'reset_state'
 ]
 
 
@@ -150,7 +151,7 @@ def _validate_command_specific_arguments(args):
     if args.command == 'init' and args.name == '*':
         raise CommandSpecificArgumentsException('You must specify a project name using the argument --name')
 
-    if args.command in ['discover_tap', 'test_tap_connection', 'run_tap', 'stop_tap', 'sync_tables']:
+    if args.command in ['discover_tap', 'test_tap_connection', 'run_tap', 'stop_tap', 'sync_tables', 'reset_state']:
         if args.tap == '*':
             raise CommandSpecificArgumentsException('You must specify a source name using the argument --tap')
         if args.target == '*':
@@ -253,6 +254,8 @@ def main():
     parser.add_argument('--force', default=False, required=False,
                         help='Force sync_tables for full sync', action='store_true'
                         )
+    parser.add_argument('--replication_method_only', default='*', type=str,
+                        help='Sync only tables which their replication method is as entered value')
 
     args = parser.parse_args()