+++ pipelinewise/cli/pipelinewise.py
@@ -8,6 +8,7 @@ import signal
 import sys
 import json
 import copy
+
 import psutil
 import pidfile
 
@@ -1397,7 +1398,9 @@ class PipelineWise:
         else:
             tables_to_sync = self.args.tables
 
-        selected_tables = self._get_sync_tables_setting_from_selection_file(tables_to_sync)
+        selected_tables = self._get_sync_tables_setting_from_selection_file(
+            tables_to_sync, self.args.replication_method_only)
+
         processes_list = []
         if selected_tables['partial_sync']:
             self._reset_state_file_for_partial_sync(selected_tables)
@@ -1861,6 +1864,34 @@ class PipelineWise:
             if cons_target_config:
                 utils.silentremove(cons_target_config)
 
+    def reset_state(self):
+        """Reset state file"""
+
+        if self.tap.get('type') == 'tap-postgres':
+            self._update_state_file('lsn', 1)
+            self.logger.info('state file is reset for log based tables!')
+        else:
+            self.logger.error('state reset is available only for PostgreSQL taps!')
+            raise SystemExit(1)
+
+    def _update_state_file(self, table_property, new_value):
+        tap_state = self.tap['files']['state']
+        try:
+            with open(tap_state, 'r', encoding='utf8') as state_file:
+                state_content = json.load(state_file)
+                bookmarks = state_content.get('bookmarks')
+                for table, properties in bookmarks.items():
+                    if table_property in properties:
+                        bookmarks[table][table_property] = new_value
+                state_content['bookmarks'] = bookmarks
+
+            with open(tap_state, 'w', encoding='utf8') as state_file:
+                json.dump(state_content, state_file, indent=4)
+
+        except Exception as exp:
+            self.logger.error(exp)
+            raise SystemExit(1) from exp
+
     @staticmethod
     def _remove_not_partial_synced_tables_from_properties(tap_params, not_synced_tables):
         """" Remove partial sync table which are not synced yet from properties """
@@ -1886,7 +1917,7 @@ class PipelineWise:
             filtered_bookmarks = dict(filter(lambda k: k[0] not in selected_partial_sync_tables, bookmarks.items()))
             state_content['bookmarks'] = filtered_bookmarks
             with open(tap_state, 'w', encoding='utf8') as state_file:
-                json.dump(state_content, state_file)
+                json.dump(state_content, state_file, indent=4)
 
     def _check_supporting_tap_and_target_for_partial_sync(self):
         tap_type = self.tap['type']
@@ -2086,7 +2117,7 @@ TAP RUN SUMMARY
                         bookmarks.pop(table_name.replace('"', ''), None)
 
                 state_file.seek(0)
-                json.dump(state_data, state_file)
+                json.dump(state_data, state_file, indent=4)
                 state_file.truncate()
 
         except FileNotFoundError:
@@ -2098,7 +2129,8 @@ TAP RUN SUMMARY
     def _get_fixed_name_of_table(stream_id):
         return stream_id.replace('-', '.', 1)
 
-    def _get_sync_tables_setting_from_selection_file(self, tables):
+    def _get_sync_tables_setting_from_selection_file(self, tables, replication_method_only='*'):
+        replication_method = replication_method_only.upper()
         selection = utils.load_json(self.tap['files']['selection'])
         selection = selection.get('selection')
         all_tables = {'full_sync': [], 'partial_sync': {}}
@@ -2107,10 +2139,11 @@ TAP RUN SUMMARY
             for table in selection:
                 table_name = self._get_fixed_name_of_table(table['tap_stream_id'])
                 if tables_list is None or table_name in tables_list:
-                    if table.get('sync_start_from'):
-                        all_tables['partial_sync'][table_name] = table['sync_start_from']
-                    else:
-                        all_tables['full_sync'].append(table_name)
+                    if replication_method in ['*', table.get('replication_method')]:
+                        if table.get('sync_start_from'):
+                            all_tables['partial_sync'][table_name] = table['sync_start_from']
+                        else:
+                            all_tables['full_sync'].append(table_name)
             return all_tables
 
     def __check_if_table_is_selected(self, table_in_properties):